+++ src/components/notfound/__tests__/index.test.tsx
@@ -9,93 +9,23 @@ jest.mock('next/router', () => ({
   })
 }))
 
-const mockFontError = jest.fn()
-const mockFontReturn = jest.fn()
-jest.mock('three/examples/jsm/loaders/FontLoader', () => ({
-  FontLoader: class {
-    load = (_: any, finish: Function, progress: Function, error: Function) => {
-      progress('progress')
-      error('error')
-      finish(mockFontReturn)
-    }
-  }
+jest.mock('@airthium/tanatloc-3d', () => ({
+  __esModule: true,
+  default: { extra: { Background: () => <div /> } }
 }))
 
-jest.mock('three/examples/jsm/geometries/TextGeometry', () => ({
-  TextGeometry: class {
-    boundingBox = {
-      min: { x: 0, y: 0, z: 0 },
-      max: { x: 1, y: 1, z: 1 }
-    }
-    computeBoundingBox = jest.fn()
-  }
-}))
-
-const mockGLTFError = jest.fn()
-const mockGLTFReturn = jest.fn()
-jest.mock('three/examples/jsm/loaders/GLTFLoader', () => ({
-  GLTFLoader: class {
-    load = (_: any, finish: Function, error: Function) => {
-      error('error')
-      finish(mockGLTFReturn())
-    }
-  }
-}))
-
-const mockIsWebGLAvailable = jest.fn()
-jest.mock('three/examples/jsm/capabilities/WebGL', () => ({
-  isWebGLAvailable: () => mockIsWebGLAvailable()
-}))
+jest.mock('next/dynamic', () => (callback: Function) => {
+  callback()
+  return () => <div />
+})
 
 Object.defineProperty(global, 'setTimeout', {
   value: (callback: Function) => callback()
 })
 
 describe('components/notfound', () => {
-  let count = 0
   beforeEach(() => {
     mockRouter.mockReset()
-
-    mockFontError.mockReset()
-    mockFontReturn.mockReset()
-
-    mockGLTFError.mockReset()
-    mockGLTFReturn.mockReset()
-    mockGLTFReturn.mockImplementation(() => ({
-      scene: {
-        children: [
-          {
-            geometry: {
-              attributes: {
-                color: {
-                  setXYZ: jest.fn
-                }
-              },
-              boundingSphere: {
-                center: {
-                  x: 0,
-                  y: 0,
-                  z: 0
-                }
-              },
-              getAttribute: () => ({
-                count: 3,
-                getY: () => {
-                  count++
-                  if (count === 1) return 450
-                  else return 750
-                },
-                setXYZ: jest.fn
-              }),
-              setAttribute: jest.fn
-            }
-          }
-        ]
-      }
-    }))
-
-    mockIsWebGLAvailable.mockReset()
-    mockIsWebGLAvailable.mockImplementation(() => true)
   })
 
   test('render', () => {
@@ -107,13 +37,6 @@ describe('components/notfound', () => {
       ]
     }
 
-    const { unmount } = render(<NotFound />)
-
-    unmount()
-  })
-
-  test('no WebGL', () => {
-    mockIsWebGLAvailable.mockImplementation(() => false)
     const { unmount } = render(<NotFound />)
 
     unmount()